tasksystem/requirements.txt
aiocontextvars==0.2.2
appdirs==1.4.3
asgiref==3.2.7
attrs==19.2.0
bcrypt==3.1.7
black==19.3b0
bottle==0.12.18
cachetools==4.0.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
common==0.1.2
contextvars==2.4
cx-Oracle==7.2.3
databases==0.2.5
Django==2.2.6
django-cors-headers==3.2.0
django-db-connection-pool==1.0.2
django-detect==1.0.16
django-dynamic-db-router==0.3.0
django-extensions==2.2.9
django-ip==1.0.2
django-localflavor==3.0.1
django-mysql==3.2.0
django-organizations==1.1.2
django-scheduler==0.9.3
django-six==1.0.4
django-tastypie==0.14.2
djangorestframework==3.10.3
docopt==0.4.0
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googlevoice==2.0.5
icalendar==4.0.5
idna==2.8
immutables==0.11
mandrill==1.0.57
MonthDelta==0.9.1
mysql-connector==2.2.9
numpy==1.17.3
oauthlib==3.1.0
Paste==3.2.5
Pillow==7.0.0
protobuf==3.11.3
psycopg2==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-mimeparse==1.6.0
python-stdnum==1.13
pytz==2019.3
rauth==0.7.3
recurly==3.3.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
toml==0.10.0
twilio==6.38.0
urllib3==1.25.6
